+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.1.3 (unreleased)
+## 0.1.3
 
 - Add `catalystcenter_device_role` resource
 - BREAKING CHANGE: Rename `catalystcenter_device` resource and data source to `catalystcenter_pnp_device`